fix:스프링 빈 순환참조 오류 방지
클래스명 변경

Related:#133
