+++ application/src/main/java/core/application/movies/repositories/comment/jpa/CommentDislikeRepositoryJPAImpl.java
@@ -1,6 +1,6 @@
 package core.application.movies.repositories.comment.jpa;
 
-import core.application.movies.repositories.comment.CommentLikeRepository;
+import core.application.movies.repositories.comment.CommentDislikeRepository;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Profile;
@@ -8,23 +8,23 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @RequiredArgsConstructor
-@Profile("!mybatis")
-public class JpaCommentLikeRepositoryImpl implements CommentLikeRepository {
+@Profile("jpa")
+public class CommentDislikeRepositoryJPAImpl implements CommentDislikeRepository {
 
-    private final JpaCommentLikeRepository jpaRepository;
+    private final JpaCommentDislikeRepository jpaRepository;
 
     @Override
-    public void saveCommentLike(Long commentId, UUID userId) {
-        jpaRepository.saveLike(commentId, userId);
+    public void saveCommentDislike(Long commentId, UUID userId) {
+        jpaRepository.saveDisLike(commentId, userId);
     }
 
     @Override
-    public boolean isExistLike(Long commentId, UUID userId) {
+    public boolean isExistDislike(Long commentId, UUID userId) {
         return jpaRepository.existsByComment_CommentIdAndUserId(commentId, userId);
     }
 
     @Override
-    public void deleteCommentLike(Long commentId, UUID userId) {
+    public void deleteCommentDislike(Long commentId, UUID userId) {
         jpaRepository.deleteByComment_CommentIdAndUserId(commentId, userId);
     }
 }
